+++ src/kandel/geometricKandel/geometricKandelInstance.ts
@@ -362,6 +362,29 @@ class GeometricKandelInstance extends CoreKandelInstance {
         params.distribution.askGives,
       );
 
+      if (!overridesWithFunds.gasLimit) {
+        const [gasEstimation, block] = await Promise.all([
+          this.geometricKandel.estimateGas.populateFromOffset(
+            firstDistribution.from,
+            firstDistribution.to,
+            params.distribution.baseQuoteTickIndex0,
+            newBaseQuoteTickOffset,
+            params.distribution.firstAskIndex,
+            rawBidGives,
+            rawAskGives,
+            rawParameters,
+            rawDepositBaseAmount,
+            rawDepositQuoteAmount,
+            overridesWithFunds,
+          ),
+          this.geometricKandel.provider.getBlock("latest"),
+        ]);
+        overridesWithFunds.gasLimit = gasEstimation.mul(120).div(100);
+        if (overridesWithFunds.gasLimit.gt(block.gasLimit)) {
+          overridesWithFunds.gasLimit = block.gasLimit;
+        }
+      }
+
       const txs = [
         await this.geometricKandel.populateFromOffset(
           firstDistribution.from,